LayoutTests/accessibility/aria-used-on-image-maps.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<img id="image" usemap="#map" src="data:image/gif;base64,R0lGODlhAQABAIAAAOTm7AAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==" width="100" height="50" alt="">
<map name="map" id="map">
    <area tabindex="0" role="button" shape="rect" coords="0,0,50,50" href="#" title="Example 1">
    <area tabindex="0" role="button" shape="rect" coords="50,0,100,50" href="#" title="Example 2">
</map>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that you can set an ARIA role on image map elements.");

    if (window.accessibilityController) {

          var group = accessibilityController.rootElement.childAtIndex(0).childAtIndex(0);
          shouldBe("group.childAtIndex(0).role", "'AXRole: AXButton'");
          shouldBe("group.childAtIndex(1).role", "'AXRole: AXButton'");
    }

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
